back-end/src/requirements.txt
Flask>=2.3.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
google-generativeai>=0.24.0
Pillow>=10.0.0
numpy>=1.24.0
pandas>=2.0.0
gunicorn>=21.0.0
